Unify package layout
