--- pymitv4/__init__.py
@@ -1,7 +1,0 @@
-"""pymitv4 is a Python package compatible with Python 3 and up.
-It can connect to Xiaomi TVs and control them."""
-
-from pymitv4.discover import Discover
-from pymitv4.control import Control
-from pymitv4.navigator import Navigator
-from pymitv4.tv import TV
