+++ setup.py
@@ -112,7 +112,8 @@ setup(
     #
     #   py_modules=["my_module"],
     #
-    packages=find_packages(),  # Required
+    packages=find_packages('src'),  # Required
+    package_dir={'': 'src'},
 
     # This field lists other packages that your project depends on to run.
     # Any package you put here will be installed by pip when your project is
