+++ src/pymitv4/__init__.py
@@ -1,7 +1,5 @@
-"""
-pymitv is a Python package compatible with version 3 and up,
-that can connect to Xiaomi TVs, and control them.
-"""
+"""pymitv4 is a Python package compatible with Python 3 and up.
+It can connect to Xiaomi TVs and control them."""
 
 
 from pymitv4.discover import Discover
